Minor items PLUS change how we call zsh...
...which will be model for how to we call any non-bash default login
shell.   That is, rather than "[shell] && exit] we do "exec [shell]"
This way, the bash shell (that calls our non-bash default login shell)
will no longer exist -- it's not needed, why keep it, it complicates the
view when using htop "tree view," it uses extra memory, and makes it
harder to follow the logic of screendoor code.